# ARROW-14026: [C++] Enable batch parallelism in Parquet scanner