This creates a new concatenating generator which converts AsyncGenerator<optional<vector<T>>> to AsyncGenerator<T> and supports async-reentrancy. When combined with a readahead generator, this enables us to get parallelism in scanning Parquet files.

This also properly chunks each row group into smaller batches as mentioned in ARROW-14024, though we can add more specific tests in that JIRA (and fix the issue for IPC as well).