@lidavidm 2021-09-20
It appears this helps a bit in some benchmarks but hurts the taxi-parquet benchmark.