@lidavidm 2021-09-22
This helps quite a bit on all the [partitioned-dataset-filter](https://conbench.ursa.dev/compare/batches/3d25d19e1b2a46468bd5ddf5e82946e2...e0751834febe4a5c93fecd597faa1986/) benchmarks.