@lidavidm 2021-09-22
I think the segfault is the same one as in https://github.com/apache/arrow/pull/11207/commits/b3e61169d621205394d28f2b6e4b515896337e5e which I've (hopefully) fixed in #11207