@pitrou 2021-10-04
@ursabot please benchmark name=file-read lang=Python